% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frm_function.R
\name{calculateTheRFM}
\alias{calculateTheRFM}
\title{RFMfunction}
\usage{
calculateTheRFM(
  transactions,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{transactions}{a data object with columns TransDate, Customer, PurchAmount, Cost}

\item{weight_recency}{a specification for the weight of the recency parameter}

\item{weight_monetary}{a specification for the weight of the monetary parameter}

\item{weight_frequency2}{a specification for the weight of the frequency parameter}
}
\value{
Data Table with RFM Scores and corresponding group \code{date} of data \code{x}
}
\description{
Function that returns the RFM Scores of a data.table.
}
\details{
\code{data} contains the transaction data. The data set must contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}
\examples{
data(transactions)
myFun(transactions, date="01.10.2020")

}
